Major update to code structure.
Changed directory structure to a more clear and concise format.

Similarly, changed code so that source files are named appropriately, and
functions are contained in structures, sort of acting like a class. Just testing
this C-class out, it seems to work ok, but I'll probably port this to C++ for
clarity.

The login prompt is contained in one source file, now. This allows for the
focusing and tab-cycling of all widgets in the login prompt container.

Authentication exec-ing, to start the user session, relies less on the xinitrc
bash script. Trying to stray away from bash and keep everything in as few
languages as possible. Final step is to modify the window manager function in
the login manager so that it returns more information such as: Name and command
to execute window manager.

That's about it! There's probably more I'm forgetting, but that's the major
stuff. Readme still needs to be updated to reflect changes, but that will come
later.